% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{buildModel}
\alias{buildModel}
\title{Model Building Function}
\usage{
buildModel(data, strNameY, pred)
}
\arguments{
\item{data}{data.frame or data.table, from which data for model is taken}

\item{strNameY}{character, name of dependent variable Y which takes the values 0 and 1}

\item{pred}{character vector, vector of predictors' names,
must be columns of \code{data}}
}
\value{
summary of model
}
\description{
Function makes model for given data table and its variables, one is name
(character) of dependent variable, other - predictors' names (as character
vector).
}
\examples{
possiblePred <- colnames(citrus)[2:6]
buildModel(citrus, "nameBin", possiblePred)
}
