% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrHeatmap.R
\name{corrHeatmap}
\alias{corrHeatmap}
\title{Visualisation of Pearson correlation coefficient}
\usage{
corrHeatmap(data, variables)
}
\arguments{
\item{data}{data.table of data we want to visualise}

\item{variables}{character vector, names of columns in \code{data}, for which
we want to compute correlation coefficient}
}
\value{
Pearson correlation coefficients heatmap
}
\description{
Function let us create table of pearson correlation coefficient between
chosen columns in data table. Displays it as a heatmap for better understanding.
}
\details{
Reminder: We can compute correlation coefficient only for numeric vectors.
}
\examples{
vars <- c("diameter", "weight", "red", "green", "blue", "nameBin")
corrHeatmap(citrus, vars)

}
