% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationCriteria.R
\name{countInfCrit}
\alias{countInfCrit}
\title{AIC/BIC for all possible models}
\usage{
countInfCrit(data = .data$citrus, strNameY = "nameBin", possPred)
}
\arguments{
\item{data}{data.frame or data.table, from which data for model is taken,
by default \code{citrus} data.table}

\item{strNameY}{character, name of dependent variable Y which takes the values 0 and 1}

\item{possPred}{character vector, vector of possible predictors' names,
must be columns of \code{data}}
}
\value{
data.table containing all possible formulas and values of AIC/BIC
for models based on them
}
\description{
This function let us count values of AIC and BIC for models created by
formulas based on all possible predictors' combinations.
}
\examples{
possiblePred <- colnames(citrus)[2:6]
countInfCrit(possiblePred)

}
