% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{CreateErrorMatrixStats}
\alias{CreateErrorMatrixStats}
\title{Create the Confusion / Error Matrix}
\usage{
CreateErrorMatrixStats(thres_prob, Y_vec, models)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than threshold, the event is counted as a success}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}

\item{models}{a fitted object or objects' list of class inheriting from "logisMod"}
}
\value{
list with Error Matrix, Accuracy of model, Sensitivity and Specificity for each
vector of transformed model's predictions (for each regression )
}
\description{
function returns list which contains Confusion / Error Matrix (i.e sum of: True Positives (TP),
False Positives (FP), True Negative (TN), False Negative (FN)), Accuracy of model,
Sensitivity and Specificity.
}
\examples{
model <- createModels(citrus, nameBin ~ diameter + green + blue)
models <- createModels(citrus, nameBin ~ diameter + green + blue, 
          nameBin ~ diameter + red + blue + green)
p <- 0.6

CreateErrorMatrixStats(p, citrus$nameBin, models)
CreateErrorMatrixStats(p, citrus$nameBin, model)

}
