% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFoldCV.R
\name{.acc}
\alias{.acc}
\title{Accuracy function}
\usage{
.acc(true, predicted, n)
}
\arguments{
\item{true}{numeric vector, true values of Y}

\item{predicted}{numeric vector, predicted values of Y}

\item{n}{size of sample}
}
\value{
accuracy, numeric
}
\description{
Accuracy function
}
\keyword{internal}
