% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{.False_Positive}
\alias{.False_Positive}
\title{Elements of Confusion / Error Matrix - False Positive}
\usage{
.False_Positive(thres_prob, resp_pred_vec, Y_vec)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than threshold, the event is counted as a success}

\item{resp_pred_vec}{numeric vector of transformed predictions (by inverse link function) of dependent variable
from data set; we recommend use function \code{\link{predictBasedData}}}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}
}
\value{
integer, a number of False Positives
}
\description{
\code{.False_Positive} returns a value of False Positives (FP) - number of bad predictions of
failures by model (actual failure was predicted by model as success)
}
\keyword{internal}
