% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationCriteria.R
\name{.getAllCombination}
\alias{.getAllCombination}
\title{All combinations of a vector}
\usage{
.getAllCombination(vec)
}
\arguments{
\item{vec}{vector of values}
}
\value{
list of combinations
}
\description{
Function gives us list of all possible combinations of a given vector
}
\examples{
.getAllCombination(c(1, 2, 3))
}
\keyword{internal}
