% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationCriteria.R
\name{.getModelFormula}
\alias{.getModelFormula}
\title{Formula for making model}
\usage{
.getModelFormula(strNameY, pred)
}
\arguments{
\item{strNameY}{character, name of dependent variable Y which takes the values 0 and 1}

\item{possPred}{character vector, vector of possible predictors' names,
must be columns of data table from which model is creating}
}
\value{
formula for model
}
\description{
Function gives us formula to create a model. Created from name of dependent
variable and list of names of possible predictors.
}
\keyword{internal}
