% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{.resPlotInside}
\alias{.resPlotInside}
\title{Plotting part of \code{resPlot}}
\usage{
.resPlotInside(data, plotData, strNameY, strNameX, resValues)
}
\arguments{
\item{data}{data.frame or data.table, from which data for model is taken}

\item{plotData}{data.table with prediction and X variable}

\item{strNameY}{character, name of dependent variable Y which takes the values 0 and 1}

\item{strNameX}{character, predictor's name, must be column of \code{data}}

\item{resValues}{numeric vector of predicted probabilities}
}
\value{
plot of probabilities
}
\description{
Responsilbe for creating plot inside of \code{resPlot} function.
}
\details{
\code{resPlot} makes data appropriate for this method.
}
\keyword{internal}
