% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFoldCV.R
\name{.splitSet}
\alias{.splitSet}
\title{Splitting function}
\usage{
.splitSet(data, K)
}
\arguments{
\item{data}{data.frame or data.table we want to split}

\item{K}{integer, number if partitions}
}
\value{
list of groups of indexes for creating subsets of given data
}
\description{
\code{.splitSet} gives partition of shuffled indexes of data table, to use
when subsets of data are needed
}
\keyword{internal}
