% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{.True_Positive}
\alias{.True_Positive}
\title{Elements of Confusion / Error Matrix - True Positive}
\usage{
.True_Positive(thres_prob, resp_pred_vec, Y_vec)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than treshold, the event is counted as a success}

\item{resp_pred_vec}{numeric vector of transformed predictions (by inverse link function) of dependent variable
from data set; output of function resp_prediction_for_model()}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}
}
\value{
integer, a number of True Positives
}
\description{
\code{.True_Positive} returns value of True Positives (TP) - number of good predictions of successes
by model (actual success was predicted by model as succes)
}
\keyword{internal}
