% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{drawROC}
\alias{drawROC}
\title{The ROC curve visualization}
\usage{
drawROC(threshold_sequance, model, strNameY)
}
\arguments{
\item{threshold_sequance}{sequance of numeric values from 0 to 1, thresholds
of success' probability - if predicted probability of dependent variable is higher than treshold, the
event is counted as a success}

\item{model}{object glm, logistic model}

\item{strNameY}{characteristic, name of dependent variable which takes values
0 and 1}
}
\value{
plot represents ROC curve
}
\description{
A receiver operating characteristic curve (ROC curve) is a graphical plot which
illustrates the performance of a binary classifier model for diffrent choices of
threshold of success' probability. Each point on curve represents False Positive
Rate (FPR) against True Positive Rate (TPR) at each threshold setting.
}
\examples{
model <- glm(nameBin ~ diameter + blue + red, data = citrus, family = binomial("logit"))
drawROC(seq(0, 1, by = 0.01), model, "nameBin") 
}
