% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{drawROCsForEachModel}
\alias{drawROCsForEachModel}
\title{The ROC curve visualization}
\usage{
drawROCsForEachModel(threshold_sequence, strNameY, models)
}
\arguments{
\item{threshold_sequence}{sequence of numeric values from 0 to 1, thresholds
of success' probability - if predicted probability of dependent variable is higher than threshold, the
event is counted as a success}

\item{strNameY}{characteristic, name of dependent variable which takes values
0 and 1}

\item{models}{a fitted object or objects' list of class inheriting from "logisMod"}
}
\value{
one plot represents ROC curves for each model
}
\description{
A receiver operating characteristic curve (ROC curve) is a graphical plot which
illustrates the performance of a binary classifier model for different choices of
threshold of success' probability. Each point on curve represents False Positive
Rate (FPR) against True Positive Rate (TPR) at each threshold setting.
}
\examples{
model1 <- createModels(citrus, nameBin ~ diameter + blue + red)
models <- createModels(citrus, nameBin ~ diameter + blue + red, nameBin ~ diameter, nameBin ~ weight + green)

drawROCsForEachModel(seq(0, 1, by = 0.05), "nameBin", model1) 
drawROCsForEachModel(seq(0, 1, by = 0.05), "nameBin", models) 

}
