% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{ElementsOfErrorMatrix}
\alias{ElementsOfErrorMatrix}
\alias{True_Positive}
\alias{False_Positive}
\alias{True_Negative}
\alias{False_Negative}
\title{Elements of Confusion / Error Matrix}
\usage{
True_Positive(thres_prob, resp_pred_vec, Y_vec)

False_Positive(thres_prob, resp_pred_vec, Y_vec)

True_Negative(thres_prob, resp_pred_vec, Y_vec)

False_Negative(thres_prob, resp_pred_vec, Y_vec)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than treshold, the event is counted as a success}

\item{resp_pred_vec}{numeric vector of transformed predictions (by inverse link function) of dependent variable
from data set; output of function resp_prediction_for_model()}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}
}
\value{
integer, sum of TP, FP, TN and FN
}
\description{
functions return elements of Error Matrix i.e sum of: True Positives (TP), False Positives (FP),
True Negative (TN), False Negative (FN)
}
\examples{
mdl <- glm(nameBin ~ diameter + green + blue, data = citrus, family = binomial("logit"))
p <- 0.6

resp_prediction_for_model <- resp_prediction_for_model(mdl)

True_Negative(thres_prob = p, resp_prediction_for_model, citrus$nameBin)
True_Positive(thres_prob = p, resp_prediction_for_model, citrus$nameBin)
False_Negative(thres_prob = p, resp_prediction_for_model, citrus$nameBin)
False_Positive(thres_prob = p, resp_prediction_for_model, citrus$nameBin)
}
