% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel.R
\name{logisMod}
\alias{logisMod}
\title{Model Building Function}
\usage{
logisMod(formula, data)
}
\arguments{
\item{formula}{an object of class "formula", a symbolic description of the
model to be fitted, same as in \link[stats]{glm}}

\item{data}{data.frame or data.table, from which data for model is taken}
}
\value{
logistic model of class glm and also new LogisMod class
}
\description{
Function makes model based on given formula, created using variable's names
from analised data table.
}
\examples{
logisMod(nameBin ~ red, citrus)
}
