% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewPredict.R
\name{predict.logisMod}
\alias{predict.logisMod}
\title{Prediction of new values}
\usage{
\method{predict}{logisMod}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "logisMod'}

\item{newdata}{data frame in which to look for variables with which to predict}

\item{...}{other parameters}
}
\description{
Prediction of new values
}
\details{
This function is fixed form of \code{stats::predict.glm}. For more details or
different version see \link[stats]{predict.glm}.
}
\examples{
library(data.table)
model <- logisMod(nameBin ~ red, citrus)
predData <- citrus[["red"]]
newData <- list()
newData[["red"]] <- seq(min(predData), max(predData), 0.1)
newData <- as.data.table(newData)

predict(model, newData)

}
