% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePredict.R
\name{resp_prediction_for_model}
\alias{resp_prediction_for_model}
\title{Transformation prediction by inverse link function of dependent variable}
\usage{
resp_prediction_for_model(model)
}
\arguments{
\item{model}{object glm, logistic model}
}
\value{
numeric vetor
}
\description{
function returns transformed prediction of dependent variable based on model and values
of predictors
}
\examples{
mdl <- glm(nameBin ~ diameter + green + blue, data = citrus, family = binomial("logit"))
summary(mdl)
head(resp_prediction_for_model(mdl))
}
