% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{resPlot.logisMod}
\alias{resPlot.logisMod}
\title{Plot of predicted probabilities}
\usage{
\method{resPlot}{logisMod}(x)
}
\arguments{
\item{x}{model of class logisMod}
}
\value{
plot of predicted probabilities for one dependent variable
}
\description{
Function plots predicted probabilities in comparison to true values. For use
when we created model and want to plot it.
}
\examples{
model <- logisMod(nameBin ~ red, citrus)
resPlot(model)
}
