% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{resPlot}
\alias{resPlot}
\title{Generic function for plotting probabilities}
\usage{
resPlot(x, ...)
}
\arguments{
\item{x}{object of proper class}

\item{...}{other parameters}
}
\description{
Generic function for plotting probabilities
}
