% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAUC.R
\name{computeAUC}
\alias{computeAUC}
\title{Computing AUC}
\usage{
computeAUC(model, strNameY)
}
\arguments{
\item{model}{glm.object, logistic model}

\item{strNameY}{characteristic, name of dependent variable which takes values
0 and 1}
}
\value{
numeric from 0 to 1, represents area under ROC curve
}
\description{
function finds the Area Under the Receiver Operating Characterisric
curve (AUC). The area under curve lies between 0 to 1. The greater value of AUC
denotes better model performance. If AUC equals 0.5, the model is random
classiffier. It means the model randomly decides that if event is success or not.
}
\examples{
model <- glm(nameBin ~ diameter + green + blue, data = citrus, family = binomial("logit"))
computeAUC(model, "nameBin")
}
