% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAUC.R
\name{computeAUC}
\alias{computeAUC}
\title{Computing AUC}
\usage{
computeAUC(strNameY, model1, ...)
}
\arguments{
\item{strNameY}{characteristic, name of dependent variable which takes values
0 and 1}

\item{model1}{glm.object, logistic model}

\item{...}{addiction logistic models; theirs AUC will be compare with the AUC for the
model1}
}
\value{
numeric from 0 to 1, represents area under ROC curve for each model
}
\description{
function finds the Area Under the Receiver Operating Characterisric
curve (AUC). The area under curve lies between 0 to 1. The greater value of AUC
denotes better model performance. If AUC equals 0.5, the model is random
classiffier. It means the model randomly decides that if event is success or not.
}
\examples{
model1 <- logisMod(nameBin ~ diameter + blue + red, data = citrus)
model2 <- logisMod(nameBin ~ weight + green + red, data = citrus)

computeAUC("nameBin", model1, model2)
}
