% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{CreateErrorMatrixStats}
\alias{CreateErrorMatrixStats}
\title{Create the Confusion / Error Matrix}
\usage{
CreateErrorMatrixStats(thres_prob, resp_pred_vec, Y_vec)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than treshold, the event is counted as a success}

\item{resp_pred_vec}{numeric vector of transformed predictions (by inverse link function) of dependent variable
from data set; output of function resp_prediction_for_model()}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}
}
\value{
list with Error Matrix, Accuracy of model, Sensitivity and Specificity.
}
\description{
functions return list which cointains Confussion / Error Matrix (i.e sum of: True Positives (TP),
False Positives (FP), True Negative (TN), False Negative (FN)), Accuracy of model,
Sensitivity and Specificity.
}
\examples{
mdl <- glm(nameBin ~ diameter + green + blue, data = citrus, family = binomial("logit"))
p <- 0.6

resp_prediction_vec <- resp_prediction_for_model(mdl)
CreateErrorMatrixStats(p, resp_prediction_vec, mdl$data$nameBin)
}
