% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevianceTest.R
\name{deviance_test}
\alias{deviance_test}
\title{Deviance test for two logistic models}
\usage{
deviance_test(model_H0, model_H1, alpha = 0.05)
}
\arguments{
\item{model_H0}{glm object, logistic model from null hypothesis (models with less
predictors are supported)}

\item{model_H1}{glm object, logistic model from alternative hypothesis (models with more
predictors are supported)}

\item{alpha}{numeric,significance level of test}
}
\value{
list of names of testing predictors, value of deviance statistic,
degrees of freedom, critical value of deviance statistic and p value.
}
\description{
function returns deviance test summary to verify null hypothesis H0: predictors
from set A is insignificant vs alternative H1: significant predictor exists in set A (full model)
}
\examples{
mdl_full <- glm(nameBin ~ diameter + green + blue + weight, 
                 data = citrus, family = binomial("logit"))
mdl_unfull <- glm(nameBin ~ weight + green, 
                 data = citrus, family = binomial("logit"))
                 
deviance_test(model_H0 = mdl_unfull, model_H1 = mdl_full, alpha = 0.1)
}
