% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCcurve.R
\name{.findPointsOnTheROC}
\alias{.findPointsOnTheROC}
\title{Finding points on the ROC curve corresponding to a chosen threshold}
\usage{
.findPointsOnTheROC(threshold_sequence, model, strNameY)
}
\arguments{
\item{threshold_sequence}{sequence of numeric values from 0 to 1, thresholds
of success' probability - if predicted probability of dependent variable is higher than threshold, the
event is counted as a success}

\item{model}{glm.object, logistic model build from data, which is data.table
object}

\item{strNameY}{characteristic, name of dependent variable which takes values
0 and 1}
}
\value{
a list of pairs represents FPR and TPR for each threshold probability
}
\description{
\code{.findPointsOnTheROC} computes a point which represents the False
Positive Rate (FPR) and the True Positive Rate (TPR, sensitivity) at each threshold
setting
}
\keyword{internal}
