% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someFunctions.R
\name{firstCase}
\alias{firstCase}
\title{sse of scaled data for linear model}
\usage{
firstCase(observations, data, parameters)
}
\arguments{
\item{observations}{numeric, vector of dependent variable}

\item{data}{numeric, plan matrix for model}

\item{parameters}{numeric, beta values}
}
\value{
numeric, sse of scaled data
}
\description{
function scales given data, then counts sse using that data and given
vector of observations
}
\examples{
n <- 88
planMatrix <- matrix(rnorm(3*n), nrow = n, ncol = 3)
beta <- c(1, 2, 1)
Y <- planMatrix \%*\% beta + rnorm(n, sd = 0.01)
firstCase(Y, planMatrix, beta)
}
