% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{resPlot.default}
\alias{resPlot.default}
\title{Plot of predicted probabilities}
\usage{
\method{resPlot}{default}(x, ...)
}
\arguments{
\item{...}{character, names of dependent variable and predictor}

\item{data}{data.frame or data.table, from which data for model is taken}
}
\value{
plot of predicted probabilities for one dependent variable
}
\description{
Function makes model for given data table and its variables, one is name
(character) of dependent variable, other - predictor's name (character). Then
plots predicted probabilities in comparison to true values.
}
\examples{
resPlot(citrus, "nameBin", "diameter")
}
