% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponsePlot.R
\name{resPlot.LogisMod}
\alias{resPlot.LogisMod}
\title{Plot of predicted probabilities}
\usage{
\method{resPlot}{LogisMod}(x)
}
\arguments{
\item{x}{model of class LogisMod}
}
\value{
plot of predicted probabilities for one dependent variable
}
\description{
Function plots predicted probabilities in comparison to true values. For use
when we created model and want to plot it.
}
\examples{
model <- buildModel(nameBin ~ red, citrus)
resPlot(model)
}
