% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizationBoxplot.R
\name{visual_boxplot}
\alias{visual_boxplot}
\title{Box plot visualization for logistic model}
\usage{
visual_boxplot(
  data,
  strNameY,
  strNameX,
  plot_title = paste("Box plot of", strNameX, "versus", strNameY)
)
}
\arguments{
\item{data}{data table or data frame, cointains names of variables}

\item{strNameY}{character, name of dependent variable Y which takes the values 0 and 1}

\item{strNameX}{character, name of independent variable/predictor X}

\item{plot_title}{character, optional, title of box plot}
}
\description{
function displays distribition data based on its five statistic:
minimum, first quartile Q1, median, third quartile Q3 and maximum
}
\details{
package patchwork is available to simplify displaying graphs
}
\examples{
citrus
str(citrus)
p1 <- visual_boxplot(data = citrus, strNameY = "nameBin", strNameX ="diameter")
p2 <- visual_boxplot(data = citrus, strNameY = "nameBin", strNameX ="green")
p1 + p2

}
