% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElementsErrorMatrix.R
\name{visualErrorMatrix}
\alias{visualErrorMatrix}
\title{Visualization of Confusion / Error Matrix}
\usage{
visualErrorMatrix(thres_prob, resp_pred_vec, Y_vec)
}
\arguments{
\item{thres_prob}{numeric from 0 to 1, threshold of success' probability - if predicted
probability of dependent variable is higher than treshold, the event is counted as a success}

\item{resp_pred_vec}{numeric vector of transformed predictions (by inverse link function) of dependent variable
from data set; output of function resp_prediction_for_model()}

\item{Y_vec}{dependent variable Y which takes the values 0 and 1}
}
\value{
plot which represents Confusion / Error Matrix
}
\description{
function displays Confusion / Error Matrix.
}
\examples{
mdl <- glm(nameBin ~ diameter + green + blue, data = citrus, family = binomial("logit"))
p <- 0.6

resp_prediction_vec <- resp_prediction_for_model(mdl)
visualErrorMatrix(p, resp_prediction_vec, mdl$data$nameBin)
}
